requirements.txt
Django>=4.0.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
torch>=2.0.0
transformers>=4.30.0
pyannote.audio>=3.0.0
SpeechRecognition>=3.10.0
pydub>=0.25.1
openai>=1.0.0
nltk>=3.8.1
scikit-learn>=1.2.2
python-multipart>=0.0.6
requests>=2.31.0 